+++ CHANGELOG.md
@@ -1,10 +1,21 @@
 #Teknoo Software - States Bundle - Change Log
 
+##[1.1.0] - 2015-12-06
+###Added
+- Stable release 1.1.0
+
+###Remove
+- Typo3 class alias loader
+
+##[1.1.0-beta6] - 2015-10-21
+###Fixed
+- Fix bug to retrieve composer instance via spl functions
+
 ##[2.0.0-beta2] - 2015-10-31 - Available on the branch "next"
 ###Changed
 - Migrate from Uni Alteri to Teknoo Software organization
 
-##[1.1.0-beta5] - 2015-07-23
+##[1.1.0-beta5] - 2015-10-31
 ###Fixed
 - Migrate from Uni Alteri to Teknoo Software organization